13.5.1-5 baseline
Former-commit-id: 492f92ef7be04103dcb14bd812af5ae8e4350b43 [formerly 492f92ef7be04103dcb14bd812af5ae8e4350b43 [formerly e0157ab61ca5f2a62d78993a27893778fa119c4f]]
Former-commit-id: ea7ac3793022325a16e6ea4db51c454f7a5e6c3b
Former-commit-id: 869922cb8275fd9eccb4277e99e0ae47afefdac6

+++ ncep/gov.noaa.nws.ncep.edex.plugin.idft/unit-test/gov/noaa/nws/ncep/edex/plugin/idft/util/IdftParserTest.java
@@ -46,7 +46,7 @@ public class IdftParserTest {
 		Matcher m2 = dataLnPattern2.matcher(thePntRec2);
 		if (m2.find()) {
 			IdftParser.processIdft(m2, 6, record);	
-		    assertEquals (410, record.getPointNum());
+		    assertEquals (410, record.getPointNum().intValue());
 			assertEquals (-66.8, record.getLat());
 			assertEquals (-73.3, record.getLon());
 			assertEquals (253.0F, record.getDirection());
@@ -56,7 +56,7 @@ public class IdftParserTest {
 		m2 = dataLnPattern2.matcher(thePntRec2);
 		if (m2.find()) {
 			IdftParser.processIdft(m2, 6, record);	
-		    assertEquals (390, record.getPointNum());
+		    assertEquals (390, record.getPointNum().intValue());
 			assertEquals (50.5, record.getLat());
 			assertEquals (10.3, record.getLon());
 			assertEquals (180.0F, record.getDirection());
